% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis-templates.R
\name{illinois_thesis_ece_pdf}
\alias{illinois_thesis_ece_pdf}
\title{Creates an R Markdown PDF Thesis document in the Style of the ECE Department at UIUC}
\usage{
illinois_thesis_ece_pdf(toc = TRUE, toc_depth = 3,
  highlight = "default", ...)
}
\arguments{
\item{toc}{A Boolean (TRUE or FALSE) specifying whether table of contents should be created}

\item{toc_depth}{A positive integer}

\item{highlight}{Syntax highlighting style. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock". Pass NULL to prevent syntax highlighting.}
}
\value{
A modified \code{pdf_document} that uses the ECE Department style at
UIUC
}
\description{
Processes the R Markdown file using the appropriate template.
}
\details{
The function is based off of \href{https://github.com/ismayc/thesisdown}{thesisdown::thesis_pdf}
format by Chester Ismay.
}
\examples{
\dontrun{
 output: uiucthemes::illinois_thesis_ece_pdf
}
}
