% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis-templates.R
\name{thesis_pdf_ece}
\alias{thesis_pdf_ece}
\title{Creates an R Markdown PDF Thesis document in the Style of the ECE Department
at UIUC}
\usage{
thesis_pdf_ece(toc_depth = 3, highlight = "default", ...)
}
\arguments{
\item{toc_depth}{Depth shown in table of contents. Must be a positive integer.}

\item{highlight}{Syntax highlighting style. Supported styles include:
"default", "tango", "pygments", "kate", "monochrome",
"espresso", "zenburn", and "haddock". Pass \code{NULL}
to prevent syntax highlighting.}

\item{...}{Additional parameters passed to \code{\link[bookdown:pdf_book]{bookdown::pdf_book()}}.}
}
\value{
A modified \code{pdf_document} that uses the ECE Department style at UIUC.
}
\description{
Processes the R Markdown file using the appropriate ECE PDF template.
}
\examples{
\dontrun{
 output: uiucthemes::thesis_pdf_ece
}
}
\references{
\itemize{
\item \href{https://github.com/ismayc/thesisdown}{thesisdown::thesis_pdf} by Chester Ismay
\itemize{
\item The \emph{R} function is largely based off of this format.
}
\item \href{https://wiki.illinois.edu/wiki/display/ECEThesisReview/LaTeX+Resources}{ECE Department Thesis LaTeX Template}
\itemize{
\item The \code{template.tex} uses the shell.
}
}
}
