% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiuc_beamer.R
\name{uiuc_beamer}
\alias{uiuc_beamer}
\title{Creates an R Markdown UIUC Beamer Themed Slide Deck}
\usage{
uiuc_beamer()
}
\value{
A modified \code{beamer_presentation} based on the UIUC Beamer Themed
  template
}
\description{
Handles routing of the template in the YML file.
}
\examples{
\dontrun{
 output:
   uiucthemes::uiuc_beamer:
     toc: true
}
}

