% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{uiucthemes-defunct}
\alias{uiucthemes-defunct}
\alias{uiuc_beamer}
\alias{uiuc_beamer}
\title{Defunct functions in the uiucthemes package}
\usage{
uiuc_beamer(...)
}
\arguments{
\item{...}{All old parameters being passed into the function.}
}
\description{
These functions have been removed from the uiucthemes package.
}
\section{Details}{

\tabular{rl}{
  \code{uiuc_beamer} \tab has been removed in favor of \code{\link{beamer_illinois}}\cr
}
}

