% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiucthemes.R
\docType{package}
\name{uiucthemes-package}
\alias{uiucthemes}
\alias{uiucthemes-package}
\title{'R' 'Markdown' Themes for 'UIUC' Documents and Presentations}
\description{
A set of custom 'R' 'Markdown' templates for documents and
presentations with the University of Illinois at Urbana-Champaign (UIUC)
color scheme and identity standards.
}
\details{
Trial run on RMarkdown templates
}

