% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiucthemes.R
\docType{package}
\name{uiucthemes-package}
\alias{uiucthemes}
\alias{uiucthemes-package}
\title{UIUC Themed RMarkdown templates}
\description{
The package contains a beamer template that generates UIUC
themed slides
}
\details{
Trial run on RMarkdown templates
}

